[cz80] Fixed the types used by the R accessor and mutator.
Cz80_Get_R(): zR only has the low 7 bits. zR2 has the high bit.
Cz80_Set_R(): Likewise.

+++ src/libgens/cz80/cz80.h
@@ -343,7 +343,7 @@ uint16_t     CZ80CALL Cz80_Get_WZ(cz80_struc *cpu);
 #endif
 uint16_t     CZ80CALL Cz80_Get_PC(cz80_struc *cpu);
 
-uint16_t     CZ80CALL Cz80_Get_R(cz80_struc *cpu);
+uint8_t      CZ80CALL Cz80_Get_R(cz80_struc *cpu);
 uint16_t     CZ80CALL Cz80_Get_IFF(cz80_struc *cpu);
 uint8_t      CZ80CALL Cz80_Get_IM(cz80_struc *cpu);
 uint8_t      CZ80CALL Cz80_Get_I(cz80_struc *cpu);
@@ -367,7 +367,7 @@ void    CZ80CALL Cz80_Set_WZ(cz80_struc *cpu, uint16_t value);
 #endif
 void    CZ80CALL Cz80_Set_PC(cz80_struc *cpu, uint16_t value);
 
-void    CZ80CALL Cz80_Set_R(cz80_struc *cpu, uint16_t value);
+void    CZ80CALL Cz80_Set_R(cz80_struc *cpu, uint8_t value);
 void    CZ80CALL Cz80_Set_IFF(cz80_struc *cpu, uint16_t value);
 void    CZ80CALL Cz80_Set_IM(cz80_struc *cpu, uint8_t value);
 void    CZ80CALL Cz80_Set_I(cz80_struc *cpu, uint8_t value);
